seperated app from server so server is not run after testing
